requirements.txt
slack-bolt==1.18.1
python-dotenv==1.0.0
openai==1.81.0
openai-agents==0.0.17
mcp-remote==0.1.0 